Move the torchvision stuff to prepare_torchvision
